// src/app/model/medico.model.ts
import { Timestamp } from "rxjs";

export class Medico {
    username: string;
    nombre: string;
    apellido: string;
    estado: string;
    email: string;
    recoveryEmail: string;
    password: string;
    descripcion: string;
    numeroContacto: string;
    profesion: string;
    slogan: string;
}